chore: seperate ui constants
